infinite scrolling in subreddit fragment. blue grey theme (only on main activity). linkify comments. hide some nav drawer items. fix wrong info shown on load issue. post age. empty search activity.

+++ app/src/main/java/com/blongdev/sift/PostInfo.java
@@ -13,7 +13,7 @@ public class PostInfo {
     public int mPoints;
     public int mComments;
     public String mUrl;
-    public int mAge;
+    public long mAge;
     public String mImageUrl;
     public boolean mFavorited;
     public String mServerId;
